bpaf_derive

Module top

Structs

Branch

EnumBranch 🔒

ParsedEnumBranch 🔒

Generating code for enum branch needs enum name which is not available from parsing at that moment so operations are performed in two steps:

Top 🔒

Top level container

Enums

Body 🔒

Describes the actual fields, can be either a single branch for struct or multiple enum variants

FieldSet 🔒

Functions

ident_to_long 🔒

ident_to_short 🔒

split_ehelp_into 🔒

split_options_help 🔒